Refactor distributed network simulator to generate one-time package associated with a simulation instance
